requirements.txt
django==2.2.3
django-extensions==2.1.9
psycopg2-binary==2.8.3

djangorestframework==3.9.4

factory-boy==2.12.0
Faker==1.0.7

django-rest-swagger==2.2.0

channels==2.2.0
channels_redis==2.4.0

ipython==7.6.0

pyfcm==1.4.7

django-cors-headers==3.1.1

celery==4.3.0
redis==3.3.11